refactor(ui): rename LoadingButton folder into SubmitButton
